backend/requirements.txt
# FastAPI and Web Server
fastapi
uvicorn[standard]

# Google GenAI SDK (New Standard - GA release as per migration guide)
# Replaces google-generativeai with improved developer experience
google-genai

# HTTP Client for GitHub API
aiohttp

# Core dependencies
python-multipart
pydantic
python-dotenv
websockets
